migrate table vdiff tests
