+++ components/table/test/screenshots/ci/golden/table/d2l-table-ltr-default-sticky-one-column.png
