+++ components/table/test/screenshots/ci/golden/table/d2l-table-ltr-light-sticky-one-column.png
