+++ components/table/test/screenshots/ci/golden/table/d2l-table-rtl-default-sticky-one-column.png
