+++ components/table/test/screenshots/ci/golden/table/d2l-table-rtl-light-sticky-one-column.png
