--- components/table/test/table.visual-diff.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<link rel="stylesheet" href="../../../test/styles.css" type="text/css">
-		<script>
-			const rtl = (window.location.search.indexOf('dir=rtl') !== -1);
-			if (rtl) document.documentElement.setAttribute('dir', 'rtl');
-		</script>
-		<script type="module">
-			import '../../typography/typography.js';
-			import './table-test-visual-diff.js';
-			import './table-test-controls-visual-diff.js';
-			import './table-test-sticky-visual-diff.js';
-			import './table-test-paging-visual-diff.js';
-		</script>
-		<title>d2l-table</title>
-		<meta name="viewport" content="width=device-width, initial-scale=1.0">
-		<meta charset="UTF-8">
-		<style>
-			html {
-				--d2l-input-position: static; /* emulates what grades is doing in the LMS */
-			}
-		</style>
-	</head>
-	<body class="d2l-typography">
-		<d2l-test-table-visual-diff></d2l-test-table-visual-diff>
-		<d2l-test-table-sticky-visual-diff></d2l-test-table-sticky-visual-diff>
-		<d2l-test-table-controls-visual-diff></d2l-test-table-controls-visual-diff>
-		<d2l-test-table-paging-visual-diff></d2l-test-table-paging-visual-diff>
-		<div style="margin-bottom: 1500px;"></div>
-	</body>
-</html>
